requirements.txt
imageio
llvmlite
numba
numpy
opencv-python
pillow
pyimgur
python-dotenv
python-telegram-bot
tqdm
